+++ install-system.bash
@@ -29,8 +29,14 @@ ODOO_HELPER_BRANCH=${1:-master}
 set -e;  # Fail on each error
 
 # Install git if not installed yet
-if ! command -v git >/dev/null 2>&1; then
-    apt-get install -yqq --no-install-recommends git wget;
+if ! command -v git >/dev/null 2>&1 || ! command -v wget >/dev/null 2>&1; then
+    if [ -e "/etc/debian_version" ]; then
+        echo -e "${BLUEC}INFO${NC}: Installing minimal system dependencies...${NC}";
+        apt-get install -yqq --no-install-recommends git wget;
+    else
+        echo -e "${REDC}ERROR${NC}: Please, install wget and git to be able to install odoo-helper-scripts!"
+        exit 2;
+    fi
 fi
 
 # define vars
@@ -49,10 +55,13 @@ ODOO_HELPER_BIN=${ODOO_HELPER_BIN:-$INSTALL_PATH/bin};
 
 # clone repo
 if [ ! -d $INSTALL_PATH ]; then
+    echo -e "${BLUEC}INFO${NC}: clonning odoo-helper-scripts...${NC}";
     git clone --recurse-submodules -q https://gitlab.com/katyukha/odoo-helper-scripts $INSTALL_PATH;
+    echo -e "${BLUEC}INFO${NC}: fetching submodules...${NC}";
     (cd $INSTALL_PATH && git checkout -q $ODOO_HELPER_BRANCH && git submodule init && git submodule update);
     # TODO: may be it is good idea to pull changes from repository if it is already exists?
     # TODO: implement here some sort of upgrade mechanism?
+    echo -e "${GREENC}OK${NC}: odoo-helper-scripts successfully clonned${NC}";
 fi
 
 # install odoo-helper user config
@@ -69,13 +78,21 @@ for oh_cmd in $ODOO_HELPER_BIN/*; do
     fi
 done
 
+
+if [ -e "/etc/debian_version" ]; then
+    odoo-helper install pre-requirements;
+    ODOO_HELPER_PRE_REQUIREMENTS_INSTALLED=1;
+fi
+
 echo -e "${YELLOWC}odoo-helper-scripts${GREENC} seems to be successfully installed system-wide!${NC}";
 echo -e "Install path is ${YELLOWC}${INSTALL_PATH}${NC}";
 echo;
-echo -e "${YELLOWC}NOTE${NC}: Do not forget to install odoo-helper system dependencies.";
-echo -e "To do this for debian-like systems run following command (${YELLOWC}sudo access required${NC}):";
-echo -e "    $ ${BLUEC}odoo-helper install pre-requirements${NC}";
-echo;
+if [ -z "$ODOO_HELPER_PRE_REQUIREMENTS_INSTALLED" ]; then
+    echo -e "${YELLOWC}NOTE${NC}: Do not forget to install odoo-helper system dependencies.";
+    echo -e "To do this for debian-like systems run following command (${YELLOWC}sudo access required${NC}):";
+    echo -e "    $ ${BLUEC}odoo-helper install pre-requirements${NC}";
+    echo;
+fi
 echo -e "${YELLOWC}NOTE2${NC}: Do not forget to install and configure postgresql.";
 echo -e "To do this for debian-like systems run following command (${YELLOWC}sudo access required${NC}):";
 echo -e "    $ ${BLUEC}odoo-helper install postgres${NC}";