+++ tests/test.bash
@@ -252,9 +252,6 @@ odoo-helper install openupgradelib
 # And show odoo-helper status after tools installed
 odoo-helper status  --tools-versions --ci-tools-versions
 
-# Install wkhtmltopdf (if it is not installed yet)
-odoo-helper install wkhtmltopdf
-
 
 echo -e "${YELLOWC}
 =================================
@@ -451,7 +448,7 @@ if python3 -c "import sys; exit(sys.version_info < (3, 9));"; then
     # Odoo 11 does not run on python 3.9, so build custom python interpreter
     odoo-install --install-dir odoo-11.0 --odoo-version 11.0 \
         --conf-opt-xmlrpc_port 8369 --conf-opt-xmlrpcs_port 8371 --conf-opt-longpolling_port 8372 \
-        --db-user odoo11 --db-pass odoo --build-python 3.8.9
+        --db-user odoo11 --db-pass odoo --build-python 3.8.13
 else
     # System python is less then 3.9, so it have to work in right way
     odoo-install --install-dir odoo-11.0 --odoo-version 11.0 \
@@ -641,7 +638,7 @@ if python3 -c "import sys; exit(sys.version_info < (3, 9));"; then
     # Odoo 11 does not run on python 3.9, so build custom python interpreter
     odoo-install --install-dir odoo-12.0 --odoo-version 12.0 \
         --conf-opt-xmlrpc_port 8369 --conf-opt-xmlrpcs_port 8371 --conf-opt-longpolling_port 8372 \
-        --db-user odoo12 --db-pass odoo --ocb --build-python 3.8.9
+        --db-user odoo12 --db-pass odoo --ocb --build-python 3.8.13
 else
     # System python is less then 3.9, so it have to work in right way
     odoo-install --install-dir odoo-12.0 --odoo-version 12.0 \
@@ -710,9 +707,7 @@ odoo-helper addons install --ual bureaucrat_helpdesk_lite;
 odoo-helper-fetch --odoo-app bureaucrat_helpdesk_lite;
 
 # Prepare to test pull updates with --do-update option
-odoo-helper fetch --oca partner-contact;
-(cd ./repositories/oca/partner-contact && git reset --hard HEAD^^^1);
-odoo-helper addons install --dir ./repositories/oca/partner-contact;
+(cd ./repositories/crnd-inc/generic-addons && git reset --hard HEAD^^^1);
 
 # Test pull-updates with --do-update option
 odoo-helper-addons pull-updates --do-update;
@@ -728,12 +723,6 @@ odoo-helper addons find-installed;
 odoo-helper-db drop odoo12-odoo-test;
 odoo-helper db drop -q odoo12-odoo-tmp;
 
-echo -e "${YELLOWC}
-=================================
-Install and check Odoo 13.0 (Py3)
-=================================
-${NC}"
-
 cd ../;
 
 
@@ -743,21 +732,27 @@ rm -rf ./odoo-10.0
 rm -rf ./odoo-11.0
 rm -rf ./odoo-12.0
 
+echo -e "${YELLOWC}
+=================================
+Install and check Odoo 13.0 (Py3)
+=================================
+${NC}"
+
 # Install odoo 13
 odoo-helper install sys-deps -y 13.0;
 odoo-helper postgres user-create odoo13 odoo;
 
-if python3 -c "import sys; exit(sys.version_info < (3, 6));"; then 
-    # Odoo 13 runs only with python 3.6+
+if python3 -c "import sys; exit(not (3, 6) <= sys.version_info <= (3, 9));"; then
+    # Odoo 13 runs only with python 3.6-3.9
     odoo-install --install-dir odoo-13.0 --odoo-version 13.0 \
         --http-port 8469 --http-host local-odoo-13 \
         --db-user odoo13 --db-pass odoo
 else
-    # System python is less then 3.6, so build python 3.7 to use for
-    # this odoo version
+    # System python is less than 3.6 or greater than 3.9,
+    # so build python 3.7 to use for this odoo version
     odoo-install --install-dir odoo-13.0 --odoo-version 13.0 \
         --http-port 8469 --http-host local-odoo-13 \
-        --db-user odoo13 --db-pass odoo --build-python 3.7.0
+        --db-user odoo13 --db-pass odoo --build-python 3.8.13
 fi
 
 cd odoo-13.0;
@@ -811,7 +806,7 @@ cd ../;
 odoo-helper install sys-deps -y 14.0;
 
 
-if python3 -c "import sys; exit(sys.version_info < (3, 6));"; then 
+if python3 -c "import sys; exit(not (3, 6) <= sys.version_info <= (3, 9));"; then
     # Odoo 14 runs only with python 3.6+
     odoo-install --install-dir odoo-14.0 --odoo-version 14.0 \
         --http-port 8569 --http-host local-odoo-14 \
@@ -822,7 +817,7 @@ else
     odoo-install --install-dir odoo-14.0 --odoo-version 14.0 \
         --http-port 8569 --http-host local-odoo-14 \
         --db-user odoo14 --db-pass odoo --create-db-user \
-        --build-python 3.7.0
+        --build-python 3.8.13
 fi
 
 cd odoo-14.0;
@@ -882,7 +877,7 @@ rm -rf ./odoo-14.0
 odoo-helper install sys-deps -y 15.0;
 
 
-if python3 -c "import sys; exit(sys.version_info < (3, 7));"; then 
+if python3 -c "import sys; exit(sys.version_info < (3, 7));"; then
     # Odoo 15 runs only with python 3.7+
     odoo-install --install-dir odoo-15.0 --odoo-version 15.0 \
         --http-port 8569 --http-host local-odoo-15 \
@@ -893,7 +888,7 @@ else
     odoo-install --install-dir odoo-15.0 --odoo-version 15.0 \
         --http-port 8569 --http-host local-odoo-15 \
         --db-user odoo15 --db-pass odoo --create-db-user \
-        --build-python 3.7.9
+        --build-python 3.8.13
 fi
 
 cd odoo-15.0;